rpcmessage: tweak the name Params to Param
The only one parameter is actually expected. It can be a complex data
type such as IMap or List combining multiple parameters but in its core
it is just a single parameter. Thus the correct naming should be just
Param.

This has mostly no real impact as it is only a convention tweak. The
existing implementations are encouraged to keep Params and switch only
with some major release to introduce API breakage. The difference is
negligible.